refactor: axios 메서드 type 지정 위치 수정 및 access token atom 삭제 [TRL-67]
