+++ src/pages/Callback/index.tsx
@@ -1,28 +1,17 @@
 import { useQuery } from '@tanstack/react-query';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
 import HTTP from '@/api';
-import AccessToken from '@/state/accessToken';
 import RedirectUrl from '@/state/redirectUrl';
-
-interface AccessTokenResponse {
-  token_type: string;
-  expires_in: number;
-  access_token: string;
-  scope: string;
-  refresh_token: string;
-}
 
 const Callback = () => {
   const [searchParams] = useSearchParams();
   const oauthCode = searchParams.get('code') || '';
-  const setAccessToken = useSetRecoilState(AccessToken);
   const redirectUrl = useRecoilValue(RedirectUrl);
   const navigate = useNavigate();
 
-  const onSuccess = (data: AccessTokenResponse) => {
-    setAccessToken(data.access_token);
+  const onSuccess = () => {
     navigate(redirectUrl);
   };
 
@@ -30,10 +19,9 @@ const Callback = () => {
     navigate('/login');
   };
 
-  // 백엔드에 get요청 보내기
-  const { data } = useQuery(
+  useQuery(
     ['getAccessToken', oauthCode],
-    () => HTTP.getAccessToken<AccessTokenResponse>(oauthCode),
+    () => HTTP.getAccessToken(oauthCode),
     {
       enabled: !!oauthCode,
       onSuccess,
