+++ src/pages/Login/OAuthBtnContainer/index.tsx
@@ -3,20 +3,10 @@ import { useQuery } from '@tanstack/react-query';
 import HTTP from '@/api';
 import OAuthBtn from '@/pages/Login/OAuthBtn';
 
-interface OAuthUriResponse {
-  id: number;
-  name: string;
-  url: string;
-}
-
 const OAuthBtnContainer = () => {
-  const { data: authUrl } = useQuery(
-    ['login-uri'],
-    () => HTTP.getLoginUri<OAuthUriResponse[]>(),
-    {
-      suspense: true,
-    }
-  );
+  const { data: authUrl } = useQuery(['login-uri'], () => HTTP.getLoginUri(), {
+    suspense: true,
+  });
 
   const OAuthBtnList = !authUrl
     ? null
